afterMessageCreated

+++ package/src/Actions/DeleteMessageAction.php
@@ -3,15 +3,29 @@
 namespace TMSPerera\HeadlessChat\Actions;
 
 use TMSPerera\HeadlessChat\Events\MessageDeletedEvent;
+use TMSPerera\HeadlessChat\Exceptions\InvalidParticipationException;
 use TMSPerera\HeadlessChat\Models\Message;
 use TMSPerera\HeadlessChat\Models\Participation;
 
 class DeleteMessageAction
 {
-    public function __invoke(Message $message, Participation $participation): void
-    {
+    /**
+     * @throws InvalidParticipationException
+     */
+    public function __invoke(
+        Message $message,
+        Participation $deleterParticipation,
+    ): void {
+        $message->loadMissing(['conversation.participations']);
+
+        $conversationParticipation = $message->conversation->participations->find($deleterParticipation);
+
+        if (! $conversationParticipation) {
+            throw new InvalidParticipationException;
+        }
+
         $message->delete();
 
-        MessageDeletedEvent::dispatch($message, $participation);
+        MessageDeletedEvent::dispatch($message, $deleterParticipation);
     }
 }
